+++ R/plotly.R
@@ -342,7 +342,7 @@ as_widget <- function(x, ...) {
   # add plotly class mainly for printing method
   # customize the JSON serializer (for htmlwidgets)
   attr(x, 'TOJSON_FUNC') <- to_JSON
-  w <- htmlwidgets::createWidget(
+  htmlwidgets::createWidget(
     name = "plotly",
     x = x,
     width = x$layout$width,
@@ -359,10 +359,6 @@ as_widget <- function(x, ...) {
       list(plotlyMainBundle())
     )
   )
-  # set an ID to avoid the rmarkdown warning ('.Random.seed' is not an integer vector but of type 'NULL', so ignored)
-  # note this will throw a warning in shiny, but it is at least less obtrusive
-  w$elementId <- w$elementId %||% new_id()
-  w
 }
 
 typedArrayPolyfill <- function() {